Fix saving functions in datapacks
Prior to 1.18.2, the server resource manager was required to be a
specific type since implementation details were used to save functions.

Commit a6c5d77b67242b441b2f256a7d78d4d7f79188fa, which updated the mod
for 1.18.2, removed the dependency on implementation details but left in
the instanceof check. This causes datapack function saving to fail in
1.19.1 since the resource manager type was changed.

This commit removes the unnecessary instanceof check and corresponding
error message.