Fix saving multiline commands in datapacks
When originally writing this code, I forgot to pack the command hash
when saving multiline commands in datapack functions. This resulted in a
malformed MessagePack file being written to disk, causing the IDE to
crash the next time it was opened.

This simply adds the missing logic to pack the command hash.