Make keyboard navigation actually usable
Previously, the command suggestor would handle all presses of the up or
down arrow keys, taking priority over the text field. Since the
suggestions window always showed whenever it could, this made vertical
keyboard navigation completely unusable.

This commit fixes that by hiding and showing the suggestions window
dynamically based on the user's previous actions. All other multiline
text editors I've used also do this.

The logic is currently to show the suggestions window if the user types
a character and hide it if the user uses a keyboard shortcut (such as
the arrow keys) or unfocuses the field.