Minor README.md adjustments for CurseForge
Ideally, I'd like to use README.md as the single source of truth for
descriptions on Modrinth and CurseForge. Since CurseForge uses HTML
instead of Markdown, Pandoc will be used to convert the file.

This commit makes some minor changes to README.md to improve the output
quality on CurseForge, at least in the WYSIWYG preview (which uses
different rendering code).