Fix missing errors in Scala IntelliJ plugin
This commit replaces the build scripts in use previously with versions
from the Fabric example mod (which are written in Groovy), slightly
modified to add Scala support and some necessary libraries. For some
reason, this fixes the broken plugin.

Also, after doing this, Java records broke (not sure why). Instead of
fixing that, this commit migrates the two records in the code to Scala
case classes.