Update to 1.20.1
The only change in this update that really affected Command Block IDE
was the switch from MatrixStack to DrawContext in the Drawable
interface, which is described on the Fabric blog.
(https://fabricmc.net/2023/05/25/120.html#screen-and-rendering-changes)

The mod will probably work with 1.20 as well, but 1.20.1 is the version
being specifically targeted.

Resolves GH-24.