Render text fields using the new texture
Minecraft 1.20.2 introduced a background texture for text field widgets.
It looks the same as the previous design, but allows for resource packs
to customize it. Unfortunately, it also makes text fields slightly
smaller than before (if you don't adjust for it).

This commit changes MultilineTextFieldWidget to use the new background
texture. It's still the wrong size; that should be resolved in the next
commit.